commit 결제 완료 페이지 구현 중

+++ src/main/java/edu/kosmo/krm/joinVO/JoinOrderHistoryVO.java
@@ -33,7 +33,7 @@ public class JoinOrderHistoryVO {
 	
 	// Member_Order Table
 	//private int order_id;
-	private int order_id;
+	private Long order_id;
 	
 	private int amount;
 	private Timestamp order_date;
